Allow mime-type to be ignored on a per method basis
This is used when you want one of your api methods to work as a standard PUT endpoint, for a file upload.

This pushes the invalid Content-Type check to past the point where we have parsed the comments in the method.

When you are using the new PassThru format you will get a magical payload param that you can use for getting the payload of the uploaded file.

This code still needs tests
We may want to make a magical ContentType param as well